+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Sleet Hello Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+
+
+
+
+
+
+## [Initial Version] - 2025-04-04
 releaseing enough features with the initial version, and I have a roadmap for future features.
 here is what this version has.
 - [x] get the greeting from hello.sleet.near/.testnet on mainnet and testnet
@@ -8,3 +14,4 @@ here is what this version has.
 - [x] near cli commands - copy them so you can run them in the terminal
 - [x] discover and open web4 sites in browser
 - [x] submit web4 apps
+- [ ] discover near dapps and open in the browser